refactor(dsg): rename http-client
