+++ packages/amplication-data-service-generator/src/util/http-client.ts
@@ -6,6 +6,11 @@ import { defaultLogger as logger } from "../server/logging";
 const httpAgent = new http.Agent({ keepAlive: true });
 const httpsAgent = new https.Agent({ keepAlive: true });
 
+const httpClient = axios.create({
+  httpAgent,
+  httpsAgent,
+});
+
 axios.interceptors.request.use((request) => {
   logger.info("Starting Request", JSON.stringify(request, null, 2));
   return request;
@@ -16,7 +21,4 @@ axios.interceptors.response.use((response) => {
   return response;
 });
 
-export const httpClient = axios.create({
-  httpAgent,
-  httpsAgent,
-});
+export { httpClient };
